requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
attrs==18.2.0
boto3==1.17.85
botocore==1.20.85
certifi==2021.5.30
dj-database-url==0.5.0
Django==3.2.3
django-debug-toolbar==3.2.1
django-heroku==0.3.1
django-storages==1.11.1
django-tex==1.1.8.post1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
pdflatex==0.1.3
Pillow==8.2.0
psycopg2==2.8.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
reportlab==3.5.67
s3transfer==0.4.2
sentry-sdk==1.1.0
six==1.16.0
sqlparse==0.4.1
tex==1.8
urllib3==1.26.5
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
